Bugfix: fix module packaging issues
A number of issues came about due to the compiler generating modules
subtly wrong. After chasing settings around for a while, I found the fix
was to specify that the output module type was "CommonJS" in
package.json and tsconfig.json.

The system is once again working in our test directory.